Edited Recieve to include clientID in zip file name and folders it extracts to

+++ Server/recieve.py
@@ -21,14 +21,15 @@ def authUser(conn):
 	authCheck = open ("/home/matrix/testing/Server/users.txt", "r+")
 	confUser = 0
 
-	while True:
+	while confUser == 0:
 		authCheck.seek(0, 0)
 		cont = 0
 		counter = 0
 		userPasses = 0
 		token = 0
 		redo = 0
-		#print ('getting user info')
+		needTok = 0
+		print ('Getting user info:')
 		data = conn.recv(1024)
 		clientID = data.decode('utf-8')
 		data = conn.recv(1024)
@@ -42,73 +43,76 @@ def authUser(conn):
 			userInfo = userInfo.rstrip('\n')
 			if counter == 0:
 				if clientID == userInfo:
-					#print ('correct user')
+					print ('\tUsername: pass')
 					cont = 1
 					counter = 1
 				else:
-					#print ('incorrect user')
+					print ('\tUsername: fail')
 					cont = 0
 					counter = 1
 			elif counter == 1:
 				if clientPWD == userInfo and cont == 1:
-					#print ('correct pass')
-					#savePos = authCheck.tell()
+					print ('\tPassword: pass')
 					cont = 2
 					counter = 2
 				else:
-					#print ('incorrect pass')
+					print ('\tPassword: fail')
 					cont = 0
 					counter = 2
 			elif counter == 2:
 				if clientTOK == userInfo and cont == 2:
-					#print ('matched token')
+					print ('\tToken: match')
 					conn.send(str.encode('welcome'))
 					confUser = 1
 					break
 				elif clientTOK == 'none' and cont == 2:
-					#print ('empty token')
+					print ('\tToken: generating')
 					conn.send(str.encode('welcome'))
 					token = random.randrange(100000, 999999)
 					conn.send(str.encode(str(token)))
 					#authCheck.seek(savePos)
 					#authCheck.write(str(token) + '\n')
 					confUser = 1
+					newTok = 1
 					break
-				elif cont == 2:
-					#print ('token mismatch')
+				else:
+					print ('\tToken: fail')
 					cont = 0
 					counter = 0
-				else:
-					print ('token failed')
 
 		if confUser == 1:
 			break
 
-	if (userPasses % 2) == 0 and (userPasses % 3) != 0:
-		if clientTOK == 'none':
-			#print ('empty token')
-			conn.send(str.encode('welcome'))
-			token = random.randrange(100000, 999999)
-			conn.send(str.encode(str(token)))
+		#print ('got here')
+		if (userPasses % 2) == 0 and (userPasses % 3) != 0:
+			if clientTOK == 'none':
+				print ('empty token')
+				conn.send(str.encode('welcome'))
+				token = random.randrange(100000, 999999)
+				conn.send(str.encode(str(token)))
+				needTok = 1
+				break
 		else:
 			conn.send(str.encode('Invalid Login'))
 			print ('sent failed message')
 
-	authCheck.seek(0, 0)
-	#print (str(userPasses))
-	while redo < userPasses:
-		authCheck.readline()
-		redo += 1
-	authCheck.truncate()
-	#authCheck.seek(userPasses, 0)
-	print (authCheck.readline().rstrip('\n') + " TEST")
-	#authCheck.seek(-1, 1)
-	authCheck.write(str(token) + '\n')
-	threaded_client(conn)
+	if needTok == 1:
+		authCheck.seek(0, 0)
+		#print (str(userPasses))
+		while redo < userPasses:
+			authCheck.readline()
+			redo += 1
+		authCheck.truncate()
+		#authCheck.seek(userPasses, 0)
+		print (authCheck.readline().rstrip('\n') + " TEST")
+		#authCheck.seek(-1, 1)
+		#print ("This is the token: " + str(token))
+		authCheck.write(str(token) + '\n')
+	threaded_client(conn, clientID)
 
-def threaded_client(conn):
+def threaded_client(conn, clientID):
 	#conn.send(str.encode("Login Succeeded"))
-	file = open("/home/matrix/testing/Server/tmp/myData.zip", "wb")
+	file = open("/home/matrix/UserStorage/" + clientID + ".zip", "wb")
 	#zipArch = ZipFile ("/home/matrix/testing/Server/tmp/myData.zip", "w")
 
 	while True:
@@ -120,14 +124,20 @@ def threaded_client(conn):
 		if not data:
 			break
 	file.close()
-	zipArch = ZipFile ("/home/matrix/testing/Server/tmp/myData.zip", "r")
-	zipArch.extractall("/home/matrix/testing/Server/clientInfo")
+
+	try:
+		zipArch = ZipFile ("/home/matrix/UserStorage/" + clientID + ".zip", "r")
+		zipArch.extractall("/home/matrix/UserStorage/" + clientID + "/")
+		zipArch.close()
+	except BadZipFile:
+		print ('Error opening zip file')
+
 	conn.close()
-	zipArch.close()
 
 while (1):
 
 	conn, addr = s.accept()
 	print ('connected to: '+addr[0]+':'+str(addr[1]))
 
-	start_new_thread(authUser,(conn,))
+	#start_new_thread(authUser,(conn,))
+	authUser(conn)
